Add files via upload
This script will introduce real LLMNR and NBT-NS requests using PowerShell on a random sample of hosts in different subnets. Any responses will be captured and logged both on the console and in the file specified by $logfile. The log will record the attacker’s IP and the request to which the attacker responded.